requirements.txt
streamlit>=1.31.0
streamlit-chat>=0.1.1
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-community>=0.0.13
langchain-huggingface>=0.0.6
faiss-cpu>=1.7.4
torch>=2.1.0
transformers>=4.36.0
pypdf>=3.17.0
sentence-transformers>=2.2.2
accelerate>=0.25.0
protobuf>=4.25.1
huggingface-hub>=0.20.1
safetensors>=0.4.1
einops>=0.7.0